pyproject.toml
[tool.poetry]
name = "ml_project"
version = "0.0.0"
description = "ML in prod HW"
authors = ["Ishkov Aleksandr <ishkov142857@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
pandas = "^1.4.2"
numpy = "^1.22.3"
scikit-learn = "^1.0.2"
click = "^7.1.2"
PyYAML = "5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "ml_project.model.fit_predict:main"